Add the "get_new_keiyoushi_urls.py" script
Keiyoushi is a very popular extension repository for all the forks of the now
dead Tachiyomi manga reader for Android. Many of the extensions available there
point to manga websites that have adult themed content, so blocking all those
sources is a good idea.

This script will fetch a copy of the index provided by Keiyoushi as well as a
copy of my hosts file and then see which sources are not already blocked and
provide a file with ready to add adresses to the host file. It was made in a
VERY quick and dirty way but it also takes little to no time to run.

Signed-off-by: ausweisnummer <163487112+ausweisnummer@users.noreply.github.com>